+++ internal/braces/braces.go
@@ -1,4 +1,4 @@
-package usecases
+package braces
 
 import (
 	"math/rand"
